Chat box now available for entire app
