+++ src/pages/HomePage/HomePage.css
@@ -2,12 +2,6 @@
     width: 70%;
     margin: 15px auto;
 }
-.chat-btn{
-    font-size: 35px;
-    position: fixed;
-    bottom: 0;
-    right: 10px;
-}
 
 .project-post{
     display: grid;
